+++ source/model.hpp
@@ -16,7 +16,7 @@ private:
 
 public:
 	Model(const Device& device);
-	Model(const Device& device, const std::vector<Vertex>& vertexData, const std::vector<uint16_t> indexData);
+	Model(const Device& device, const std::vector<Vertex>& vertexData, const std::vector<uint32_t> indexData);
 
 	~Model();
